Code refactoring + doc
