+++ include/roboptim/core/manifold-map/decorator/manifold-desc.hh
@@ -36,8 +36,8 @@ namespace roboptim
   template<template <typename> class ... Types>
   struct ManiDesc
   {
-    /// \brief creates the actual manifold from the underlying library, given
-    /// a function
+    /// \brief creates the actual const manifold from the underlying library,
+    /// given a function.
     ///
     /// \tparam U the function type
     ///
@@ -50,6 +50,6 @@ namespace roboptim
   /// @}
 }
 
-# include "manifold-desc.hxx"
+# include <roboptim/core/manifold-map/decorator/manifold-desc.hxx>
 
 #endif //! ROBOPTIM_CORE_MANIFOLD_MAP_DECORATOR_MANIFOLD_DESC_HH
